# server_2/api/api.yml
swagger: "2.0"
info:
  description: "This is the facteurs-charge website's API."
  version: "1.0.0"
  title: "Facteurs-charge"
  license:
    name: "GPL-3.0"
    url: "https://www.gnu.org/licenses/gpl-3.0.fr.html"
host: "www.facteurs-charge.fr"
basePath: "/api/v1"
...
schemes:
- "https"
paths:
...
  /zones/installations:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve installations
      operationId: get_installations
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/installationSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/installationSnapshotZone'
  /zones/installations/breakdown:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the breadkdowns
      operationId: get_breakdowns
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/breakdownSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/breakdownSnapshotZone'
  /zones/installations/{installationType}:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the specified installations
      operationId: get_installation
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/installationType'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/installationSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/installationSnapshotZone'
  /zones/installations/load:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the load of the installations
      operationId: get_installations_load
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshotZone'
  /zones/installations/load/last:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the load of the installations
      operationId: get_installations_load_last
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/filter'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadFilterSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshotZone'
  /zones/installations/load/breakdown:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the load breakdown of the installations
      operationId: get_installations_load_breakdown
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadBreakdownSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadBreakdownSnapshotZone'
  /zones/installations/load/{installationType}:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the load of the specified installation
      operationId: get_installation_load
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/installationType'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshotZone'
  /zones/installations/production:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the production of the installations
      operationId: get_installations_production
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshotZone'
  /zones/installations/production/breakdown:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the production breakdown of the installations
      operationId: get_installations_production_breakdown
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerBreakdownSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerBreakdownSnapshotZone'
  /zones/installations/production/{installationType}:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the production of the specified installation
      operationId: get_installation_production
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/installationType'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshotZone'
  /zones/installations/capacity:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the capacity of the installations
      operationId: get_installations_capacity
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshotZone'
  /zones/installations/capacity/breakdown:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the capacity breakdown of the installations
      operationId: get_installations_capacity_breakdown
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerBreakdownSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerBreakdownSnapshotZone'
  /zones/installations/capacity/{installationType}:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the capacity of the specified installation
      operationId: get_installation_capacity
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/installationType'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshotZone'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshotZone'
  /zones/{zoneId}/installations:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve installation for this zone
      operationId: get_zone_installations
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/installationSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/installationSnapshot'
  /zones/{zoneId}/installations/breakdown:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the breadkdown for this zone
      operationId: get_zone_breakdown
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/breakdownSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/breakdownSnapshot'
  /zones/{zoneId}/installations/{installationType}:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the specified installation for this zone
      operationId: get_zone_installation
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/installationType'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/installationSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/installationSnapshot'
  /zones/{zoneId}/installations/load:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the load of the installations for this zone
      operationId: get_zone_installations_load
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshot'
  /zones/{zoneId}/installations/load/breakdown:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the load breakdown of the installations for this zone
      operationId: get_zone_installations_load_breakdown
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadBreakdownSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadBreakdownSnapshot'
  /zones/{zoneId}/installations/load/{installationType}:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the load of the specified installation for this zone
      operationId: get_zone_installation_load
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/installationType'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/loadSnapshot'
  /zones/{zoneId}/installations/production:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the production of the installations for this zone
      operationId: get_zone_installations_production
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshot'
  /zones/{zoneId}/installations/production/breakdown:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the production breakdown of the installations for this zone
      operationId: get_zone_installations_production_breakdown
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerBreakdownSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerBreakdownSnapshot'
  /zones/{zoneId}/installations/production/{installationType}:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the production of the specified installation for this zone
      operationId: get_zone_installation_production
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/installationType'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshot'
  /zones/{zoneId}/installations/capacity:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the capacity of the installations for this zone
      operationId: get_zone_installations_capacity
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerSnapshot'
  /zones/{zoneId}/installations/capacity/breakdown:
    x-swagger-router-controller: "Installation"
    get:
      tags:
      - Installation
      summary: Retrieve the capacity breakdown of the installations for this zone
      operationId: get_zone_installations_capacity_breakdown
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - $ref: '#/parameters/zoneId'
        - $ref: '#/parameters/from'
        - $ref: '#/parameters/to'
      responses:
        "200":
          description: "All data retreived"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerBreakdownSnapshot'
        "400":
          description: "Bad parameter"
        "404":
          description: "No data"
        "409":
          description: "Partial content"
          schema:
            type: array
            items:
              $ref: '#/definitions/powerBreakdownSnapshot'
...
definitions:
...
  breakdownSnapshotZone:
    type: object
    required: ['zoneId', 'snapshots']
    properties:
      zoneId:
        type: string
        pattern: "[A-Z]{2}(-[0-9]{2})?$"
      snapshots:
        type: array
        items:
          $ref: '#/definitions/breakdownSnapshot'
...
  loadBreakdownSnapshotZone:
    type: object
    required: ['zoneId', 'snapshots']
    properties:
      zoneId:
        type: string
        pattern: "[A-Z]{2}(-[0-9]{2})?$"
      snapshots:
        type: array
        items:
          $ref: '#/definitions/loadBreakdownSnapshot'
...
  installationSnapshotZone:
    type: object
    required: ['zoneId', 'snapshots']
    properties:
      zoneId:
        type: string
        pattern: "[A-Z]{2}(-[0-9]{2})?$"
      snapshots:
        type: array
        items:
          $ref: '#/definitions/installationSnapshot'
...
  loadSnapshotZone:
    type: object
    required: ['zoneId', 'snapshots']
    properties:
      zoneId:
        type: string
        pattern: "[A-Z]{2}(-[0-9]{2})?$"
      snapshots:
        type: array
        items:
          $ref: '#/definitions/loadSnapshot'
  loadFilterSnapshotZone:
    type: object
    required: ['zoneId', 'snapshots']
    properties:
      zoneId:
        type: string
        pattern: "[A-Z]{2}(-[0-9]{2})?$"
      snapshots:
        type: array
        items:
          $ref: '#/definitions/loadFilterSnapshot'
...
parameters:
  from:
    in: "query"
    name: "from"
    description: "Start of the period"
    required: false
    type: "string"
    format: date-time
  to:
    in: "query"
    name: "to"
    description: "End of the period"
    required: false
    type: "string"
    format: date-time
...
  installationType:
    in: "path"
    name: "installationType"
    description: "Installation type"
    required: true
    type: string
    pattern: "solar|wind|hydraulic|nuclear|bioenergy|fossil"
  filter:
    in: "query"
    name: "filter"
    description: "Highest / Lowest"
    required: false
    type: string
    pattern: "lowest|highest"